refactor: move FAQs to shared component
